Move default search entities to own modules, hook into action
Adapt tests
