+++ ckanext/search/entities/__init__.py
@@ -1,0 +1,2 @@
+from .dataset import DatasetSearchEntity
+from .organization import OrganizationSearchEntity
